backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
bcrypt
pyjwt
pydantic[email]
httpx